requirements.txt
psycopg2-binary
pythonping
mariadb
mysql-connector-python
pyodbc